refactor(tests): configure URLs in test/HttpRedirectionTest.yml file
