+++ test/HttpRedirectionTest.yml
@@ -1,0 +1,45 @@
+https://live.docs.camunda.org:
+  /:
+    target: /manual/7.10
+    status: 301
+https://stage.docs.camunda.org:
+  /:
+    target: /manual/latest/
+    status: 301
+  /optimize:
+    target: /optimize/latest/
+    status: 301
+  /optimize/user-guide/homepage:
+    target: /optimize/develop/user-guide/homepage/
+    status: 301
+  /manual/reference/bpmn20:
+    target: /manual/develop/reference/bpmn20
+    status: 301
+  /get-started/cmmn12:
+    target: /get-started/cmmn11
+    status: 307
+  /get-started/cmmn10/project-setup:
+    target: /get-started/cmmn11/project-setup
+    status: 307
+  /optimize/2.5/user-guide/homepage:
+    target: /optimize/2.5/user-guide/homepage
+    status: 301
+  /optimize/develop/technical-guide/configuration:
+    target: /optimize/develop/technical-guide/setup/configuration
+    status: 301
+  /optimize/latest/technical-guide/configuration:
+    target: /optimize/latest/technical-guide/setup/configuration
+    status: 301
+  /optimize/develop/technical-guide/installation:
+    target: /optimize/develop/technical-guide/setup/installation
+    status: 301
+  /optimize/develop/technical-guide/installation:
+    target: /optimize/latest/technical-guide/setup/installation
+    status: 301
+  /cawemo:
+    target: /cawemo/latest
+    status: 307
+  /cawemo/1.0:
+    target: /cawemo/develop
+    status: 307
+
